+++ src/MessageHandler/UpsertProjectHandler.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace App\MessageHandler;
+
+use App\Message\UpsertProjectMessage;
+use App\Service\DataProviderService;
+use Psr\Log\LoggerInterface;
+use Symfony\Component\Messenger\Attribute\AsMessageHandler;
+use Symfony\Component\Messenger\Exception\UnrecoverableMessageHandlingException;
+
+#[AsMessageHandler]
+readonly class UpsertProjectHandler
+{
+    public function __construct(
+        private LoggerInterface $logger,
+        private DataProviderService $dataProviderService,
+    ) {
+    }
+
+    public function __invoke(UpsertProjectMessage $message): void
+    {
+        try {
+            $this->logger->info('Upserting project: '.$message->projectData->name);
+            $this->dataProviderService->upsertProject($message->projectData);
+        } catch (\Exception $e) {
+            $this->logger->error($e->getMessage());
+            throw new UnrecoverableMessageHandlingException($e->getMessage());
+        }
+    }
+}