+++ src/MessageHandler/UpsertWorkerHandler.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace App\MessageHandler;
+
+use App\Message\UpsertWorkerMessage;
+use App\Service\DataProviderService;
+use Psr\Log\LoggerInterface;
+use Symfony\Component\Messenger\Attribute\AsMessageHandler;
+use Symfony\Component\Messenger\Exception\UnrecoverableMessageHandlingException;
+
+#[AsMessageHandler]
+readonly class UpsertWorkerHandler
+{
+    public function __construct(
+        private LoggerInterface $logger,
+        private DataProviderService $dataProviderService,
+    ) {
+    }
+
+    public function __invoke(UpsertWorkerMessage $message): void
+    {
+        try {
+            $this->logger->info('Upserting worker: '.$message->workerData->email);
+            $this->dataProviderService->upsertWorker($message->workerData);
+        } catch (\Exception $e) {
+            $this->logger->error($e->getMessage());
+            throw new UnrecoverableMessageHandlingException($e->getMessage());
+        }
+    }
+}